# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Dokumentasi API D4 TRPL 2019 TA13.",
        "title": "API TA",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:6900",
    "basePath": "/",
    "paths": {
        "/alternative": {
            "get": {
                "description": "Get All Alternatives",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "alternative"
                ],
                "summary": "Get All Alternatives",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/dto.AlternativesGetResponseDoc"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/response.errorResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/response.errorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/response.errorResponse"
                        }
                    }
                }
            },
            "post": {
                "description": "Create Alternative",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "alternative"
                ],
                "summary": "Create Alternative",
                "parameters": [
                    {
                        "description": "request body",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dto.AlternativeCreateRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/dto.AlternativeCreateResponseDoc"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/response.errorResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/response.errorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/response.errorResponse"
                        }
                    }
                }
            }
        },
        "/alternative/{collection_id}": {
            "get": {
                "description": "Get Alternatives By Collection ID",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "alternative"
                ],
                "summary": "Get Alternatives By Collection ID",
                "parameters": [
                    {
                        "type": "string",
                        "description": "id path",
                        "name": "collection_id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/dto.AlternativeGetByCollectionIDResponseDoc"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/response.errorResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/response.errorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/response.errorResponse"
                        }
                    }
                }
            }
        },
        "/alternative/{id}": {
            "delete": {
                "description": "Delete alternative",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "alternative"
                ],
                "summary": "Delete alternative",
                "parameters": [
                    {
                        "type": "string",
                        "description": "id path",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/dto.AlternativeDeleteResponseDoc"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/response.errorResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/response.errorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/response.errorResponse"
                        }
                    }
                }
            },
            "patch": {
                "description": "Update alternative",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "alternative"
                ],
                "summary": "Update alternative",
                "parameters": [
                    {
                        "type": "string",
                        "description": "id path",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "request body",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dto.AlternativeUpdateRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/dto.AlternativeUpdateResponseDoc"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/response.errorResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/response.errorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/response.errorResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "abstraction.PaginationInfo": {
            "type": "object",
            "properties": {
                "count": {
                    "type": "integer"
                },
                "more_records": {
                    "type": "boolean"
                },
                "page": {
                    "type": "integer"
                },
                "page_size": {
                    "type": "integer"
                },
                "sort": {
                    "type": "string"
                },
                "sort_by": {
                    "type": "string"
                }
            }
        },
        "dto.AlternativeCreateRequest": {
            "type": "object",
            "properties": {
                "Aksesibilitas": {
                    "type": "string"
                },
                "collection_id": {
                    "type": "string"
                },
                "estetika": {
                    "type": "string"
                },
                "jarak-tps": {
                    "type": "string"
                },
                "jarak_tpa": {
                    "type": "string"
                },
                "jumlah_penduduk": {
                    "type": "string"
                },
                "lokasi": {
                    "type": "string"
                },
                "nama": {
                    "type": "string"
                },
                "timbulan_sampah": {
                    "type": "string"
                }
            }
        },
        "dto.AlternativeCreateResponse": {
            "type": "object",
            "properties": {
                "Aksesibilitas": {
                    "type": "string"
                },
                "collection_id": {
                    "type": "string"
                },
                "created_at": {
                    "type": "string"
                },
                "created_by": {
                    "type": "string"
                },
                "estetika": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "jarak-tps": {
                    "type": "string"
                },
                "jarak_tpa": {
                    "type": "string"
                },
                "jumlah_penduduk": {
                    "type": "string"
                },
                "lokasi": {
                    "type": "string"
                },
                "modified_at": {
                    "type": "string"
                },
                "modified_by": {
                    "type": "string"
                },
                "nama": {
                    "type": "string"
                },
                "timbulan_sampah": {
                    "type": "string"
                }
            }
        },
        "dto.AlternativeCreateResponseDoc": {
            "type": "object",
            "properties": {
                "body": {
                    "type": "object",
                    "properties": {
                        "data": {
                            "$ref": "#/definitions/dto.AlternativeCreateResponse"
                        },
                        "meta": {
                            "$ref": "#/definitions/response.Meta"
                        }
                    }
                }
            }
        },
        "dto.AlternativeDeleteResponse": {
            "type": "object",
            "properties": {
                "Aksesibilitas": {
                    "type": "string"
                },
                "collection_id": {
                    "type": "string"
                },
                "created_at": {
                    "type": "string"
                },
                "created_by": {
                    "type": "string"
                },
                "estetika": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "jarak-tps": {
                    "type": "string"
                },
                "jarak_tpa": {
                    "type": "string"
                },
                "jumlah_penduduk": {
                    "type": "string"
                },
                "lokasi": {
                    "type": "string"
                },
                "modified_at": {
                    "type": "string"
                },
                "modified_by": {
                    "type": "string"
                },
                "nama": {
                    "type": "string"
                },
                "timbulan_sampah": {
                    "type": "string"
                }
            }
        },
        "dto.AlternativeDeleteResponseDoc": {
            "type": "object",
            "properties": {
                "body": {
                    "type": "object",
                    "properties": {
                        "data": {
                            "$ref": "#/definitions/dto.AlternativeDeleteResponse"
                        },
                        "meta": {
                            "$ref": "#/definitions/response.Meta"
                        }
                    }
                }
            }
        },
        "dto.AlternativeGetByCollectionIDResponse": {
            "type": "object",
            "properties": {
                "datas": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/entity.AlternativeEntityModel"
                    }
                }
            }
        },
        "dto.AlternativeGetByCollectionIDResponseDoc": {
            "type": "object",
            "properties": {
                "body": {
                    "type": "object",
                    "properties": {
                        "data": {
                            "$ref": "#/definitions/dto.AlternativeGetByCollectionIDResponse"
                        },
                        "meta": {
                            "$ref": "#/definitions/response.Meta"
                        }
                    }
                }
            }
        },
        "dto.AlternativeUpdateRequest": {
            "type": "object",
            "required": [
                "id"
            ],
            "properties": {
                "Aksesibilitas": {
                    "type": "string"
                },
                "collection_id": {
                    "type": "string"
                },
                "created_at": {
                    "type": "string"
                },
                "created_by": {
                    "type": "string"
                },
                "estetika": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "jarak-tps": {
                    "type": "string"
                },
                "jarak_tpa": {
                    "type": "string"
                },
                "jumlah_penduduk": {
                    "type": "string"
                },
                "lokasi": {
                    "type": "string"
                },
                "modified_at": {
                    "type": "string"
                },
                "modified_by": {
                    "type": "string"
                },
                "nama": {
                    "type": "string"
                },
                "timbulan_sampah": {
                    "type": "string"
                }
            }
        },
        "dto.AlternativeUpdateResponse": {
            "type": "object",
            "properties": {
                "Aksesibilitas": {
                    "type": "string"
                },
                "collection_id": {
                    "type": "string"
                },
                "created_at": {
                    "type": "string"
                },
                "created_by": {
                    "type": "string"
                },
                "estetika": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "jarak-tps": {
                    "type": "string"
                },
                "jarak_tpa": {
                    "type": "string"
                },
                "jumlah_penduduk": {
                    "type": "string"
                },
                "lokasi": {
                    "type": "string"
                },
                "modified_at": {
                    "type": "string"
                },
                "modified_by": {
                    "type": "string"
                },
                "nama": {
                    "type": "string"
                },
                "timbulan_sampah": {
                    "type": "string"
                }
            }
        },
        "dto.AlternativeUpdateResponseDoc": {
            "type": "object",
            "properties": {
                "body": {
                    "type": "object",
                    "properties": {
                        "data": {
                            "$ref": "#/definitions/dto.AlternativeUpdateResponse"
                        },
                        "meta": {
                            "$ref": "#/definitions/response.Meta"
                        }
                    }
                }
            }
        },
        "dto.AlternativesGetResponse": {
            "type": "object",
            "properties": {
                "datas": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/entity.AlternativeEntityModel"
                    }
                }
            }
        },
        "dto.AlternativesGetResponseDoc": {
            "type": "object",
            "properties": {
                "body": {
                    "type": "object",
                    "properties": {
                        "data": {
                            "$ref": "#/definitions/dto.AlternativesGetResponse"
                        },
                        "meta": {
                            "$ref": "#/definitions/response.Meta"
                        }
                    }
                }
            }
        },
...
        "entity.AlternativeEntityModel": {
            "type": "object",
            "properties": {
                "Aksesibilitas": {
                    "type": "string"
                },
                "collection_id": {
                    "type": "string"
                },
                "created_at": {
                    "type": "string"
                },
                "created_by": {
                    "type": "string"
                },
                "estetika": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "jarak-tps": {
                    "type": "string"
                },
                "jarak_tpa": {
                    "type": "string"
                },
                "jumlah_penduduk": {
                    "type": "string"
                },
                "lokasi": {
                    "type": "string"
                },
                "modified_at": {
                    "type": "string"
                },
                "modified_by": {
                    "type": "string"
                },
                "nama": {
                    "type": "string"
                },
                "timbulan_sampah": {
                    "type": "string"
                }
            }
        },
...
        "response.Meta": {
            "type": "object",
            "properties": {
                "info": {
                    "$ref": "#/definitions/abstraction.PaginationInfo"
                },
                "message": {
                    "type": "string",
                    "default": "true"
                },
                "success": {
                    "type": "boolean",
                    "default": true
                }
            }
        },
        "response.errorResponse": {
            "type": "object",
            "properties": {
                "error": {
                    "type": "string"
                },
                "meta": {
                    "$ref": "#/definitions/response.Meta"
                }
            }
        }
    }
}